Remove unnecessary casts reported by Eclipse
Removed about 100 casts that still remained in older Mapyrus
code that was written before Java Generics was available.

+++ src/main/java/org/mapyrus/Constants.java
@@ -194,7 +194,7 @@ public class Constants
 	{
 		int resolution = getScreenResolution();
 		double width = Toolkit.getDefaultToolkit().getScreenSize().getWidth();
-		return(width / (double)resolution * MM_PER_INCH);
+		return(width / resolution * MM_PER_INCH);
 	}
 
 	/**
@@ -205,7 +205,7 @@ public class Constants
 	{
 		int resolution = getScreenResolution();
 		double height = Toolkit.getDefaultToolkit().getScreenSize().getHeight();
-		return(height / (double)resolution * MM_PER_INCH);
+		return(height / resolution * MM_PER_INCH);
 	}
 
 	/**
